add cron to driver loading and refactor for git diff
