+++ docker-compose.yml
@@ -11,8 +11,8 @@ services:
     # Docker exposes these ports to all containers in the compose by default
     # but if you are debugging on a local machine uncomment to test connectivity or open a connection
     # to the Postgres database
-    #    ports:
-    #      - "5432:5432"
+    ports:
+      - "5432:5432"
     healthcheck:
       test: [ "CMD-SHELL", "pg_isready -U postgres" ]
       interval: 10s
@@ -55,7 +55,7 @@ services:
     ports:
       # Set port as an environment variable
       # The e2e-test forcefully sets this itself otherwise set via command line
-      - "${API_PORT}:8443"
+      - "8443:8443"
     healthcheck:
       test: [ "CMD-SHELL", "curl -X GET http://localhost:8443/health --insecure" ]
       interval: 5s
@@ -100,6 +100,8 @@ services:
       - AB2D_KEYSTORE_LOCATION=${AB2D_KEYSTORE_LOCATION}
       - AB2D_KEYSTORE_PASSWORD=${AB2D_KEYSTORE_PASSWORD}
       - AB2D_KEY_ALIAS=${AB2D_KEY_ALIAS}
+      # Update coverage frequently so that tests aren't slowed down too much
+      - "COVERAGE_UPDATE_LOAD_SCHEDULE=0/1 * * * * ?"
     depends_on:
       db:
         condition: service_healthy
