+++ worker/src/main/java/gov/cms/ab2d/worker/processor/ContractData.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.worker.processor.eob;
+package gov.cms.ab2d.worker.processor;
 
 import gov.cms.ab2d.common.model.Contract;
 import lombok.AllArgsConstructor;
