+++ worker/src/main/java/gov/cms/ab2d/worker/processor/JobProcessorImpl.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.worker.processor.eob;
+package gov.cms.ab2d.worker.processor;
 
 import com.newrelic.api.agent.NewRelic;
 import com.newrelic.api.agent.Segment;
