+++ worker/src/main/java/gov/cms/ab2d/worker/processor/PatientClaimsRequest.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.worker.processor.eob;
+package gov.cms.ab2d.worker.processor;
 
 import com.newrelic.api.agent.Token;
 import gov.cms.ab2d.common.model.CoverageSummary;
