+++ worker/src/main/java/gov/cms/ab2d/worker/service/WorkerServiceImpl.java
@@ -4,8 +4,8 @@ import gov.cms.ab2d.common.model.Job;
 import gov.cms.ab2d.common.service.PropertiesService;
 import gov.cms.ab2d.common.service.FeatureEngagement;
 import gov.cms.ab2d.common.util.Constants;
-import gov.cms.ab2d.worker.processor.eob.JobPreProcessor;
-import gov.cms.ab2d.worker.processor.eob.JobProcessor;
+import gov.cms.ab2d.worker.processor.JobPreProcessor;
+import gov.cms.ab2d.worker.processor.JobProcessor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
