+++ worker/src/test/java/gov/cms/ab2d/worker/processor/ContractProcessorImplTest.java
@@ -1,10 +1,12 @@
-package gov.cms.ab2d.worker.processor.eob;
+package gov.cms.ab2d.worker.processor;
 
 import ca.uhn.fhir.context.FhirContext;
 import gov.cms.ab2d.common.model.CoverageSummary;
 import gov.cms.ab2d.common.model.Identifiers;
 import gov.cms.ab2d.common.repository.JobRepository;
 import gov.cms.ab2d.eventlogger.LogManager;
+import gov.cms.ab2d.worker.processor.ContractProcessorImpl;
+import gov.cms.ab2d.worker.processor.PatientClaimsProcessor;
 import gov.cms.ab2d.worker.service.FileService;
 import org.hl7.fhir.dstu3.model.*;
 import org.junit.jupiter.api.BeforeEach;
@@ -14,7 +16,7 @@ import org.mockito.Mock;
 
 import java.util.*;
 
-import static gov.cms.ab2d.worker.processor.eob.ContractProcessorImpl.ID_EXT;
+import static gov.cms.ab2d.worker.processor.ContractProcessorImpl.ID_EXT;
 import static gov.cms.ab2d.worker.processor.coverage.CoverageMappingCallable.*;
 import static java.util.Collections.singletonList;
 import static org.junit.jupiter.api.Assertions.*;
