+++ worker/src/test/resources/application.properties
@@ -48,6 +48,7 @@ coverage.update.stale.days=7
 coverage.update.max.attempts=3
 coverage.update.stuck.hours=24
 coverage.update.initial.delay=1
+coverage.update.load.schedule=0 0/1 * * * ?
 
 ## ----------------------------------------------------------------------------- STUCK JOB
 ## -- run every 2 hours
